[feat] Implement barriers logic
Not final implementation! Now, the snake behaviour with the barriers is
fine: it dies on crash. But there is a problem with the coordinates
generation. It is possible that apple, barrier and snake born places
overlaps. Lists should be purged.

On the other hand the current barrier coordinates generation is very
basic. It should be revised. Random is not a good alternative because
there are positions that should not be barriers, like map confines.